requirements.txt
python-dotenv
pandas
pytest
pytest-cov
colorama
